Portal destruction in room state changes
Now you can customize whenever if it gets destroyed when room is empty, and if so, in how much seconds.

+++ src/server/websocket/index.ts
@@ -8,6 +8,7 @@ import WSMessage from './models/message'
 
 import log from '../../utils/log.utils'
 import client, { createPubSubClient } from '../../config/redis.config'
+import config from '../../config/defaults'
 
 import handleMessage from './handlers/message'
 import { PortalEvent } from '../../models/portal/defs'
@@ -112,14 +113,16 @@ export default (wss: Server) => {
                 if(extractUserId(room.controller) === socket.user.id)
                     room.releaseControl(socket.user)
 
-                setTimeout(async () =>
-                    (await room.load(room.id)).fetchOnlineMemberIds().then(({ portal, online }) => {
-                        if(online.length > 0) return
-                        if(UNALLOCATED_PORTALS_KEYS.indexOf(portal.status) > -1) return
+                if (config.destroy_portal_when_empty) {
+                    setTimeout(async () =>
+                        (await room.load(room.id)).fetchOnlineMemberIds().then(({ portal, online }) => {
+                            if(online.length > 0) return
+                            if(UNALLOCATED_PORTALS_KEYS.indexOf(portal.status) > -1) return
 
-                        room.destroyPortal()
-                    }).catch(console.error)
-                , 1000 * 5)
+                            room.destroyPortal()
+                        }).catch(console.error), config.empty_room_portal_destroy * 1000
+                    )
+                }
             }
         })
     })
